+++ bothub/api/v2/zeroshot/views.py
@@ -1,3 +1,8 @@
+import requests
+import logging
+
+from django.conf import settings
+
 from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework import permissions
@@ -9,6 +14,10 @@ from bothub.common.models import (
     Repository,
 )
 
+from bothub.api.v2.zeroshot.permissions import ZeroshotTokenPermission
+
+
+logger = logging.getLogger(__name__)
 
 class ZeroShotOptionsTextAPIView(APIView):
 
@@ -63,3 +72,44 @@ class ZeroShotRepositoryAPIView(APIView):
             "created_at": repo.created_at,
         }
         return Response(status=200, data=data)
+
+
+class ZeroShotFastPredictAPIView(APIView):
+
+    permission_classes = [ZeroshotTokenPermission]
+
+    def post(self, request):
+        data = request.data
+
+        classes = {}
+
+        for categorie in data.get("categories"):
+            option = categorie.get("option")
+            classes[option] = [option]
+            for synonym in categorie.get("synonyms"):
+                classes[option].append(synonym)
+
+        body = {
+            "input": {
+                "text": data.get("text"),
+                "classes": classes
+            }
+        }
+
+        headers = {
+            "Content-Type": "application/json; charset: utf-8",
+            "Authorization": f"Bearer {settings.ZEROSHOT_TOKEN}",
+        }
+
+        try:
+            url = settings.ZEROSHOT_BASE_NLP_URL
+            if len(settings.ZEROSHOT_SUFFIX) > 0:
+                url += settings.ZEROSHOT_SUFFIX
+            response_nlp = requests.post(
+                headers=headers,
+                url=url,
+                json=body
+            )
+        except Exception as error:
+            logger.error(f"[ - ] Zeroshot fast predict: {error}")
+        return Response(status=response_nlp.status_code, data=response_nlp.json() if response_nlp.status_code == 200 else {"error": response_nlp.text})