+++ bothub/project/usecases/project/projectdto.py
@@ -1,0 +1,12 @@
+from dataclasses import dataclass
+
+
+@dataclass
+class ProjectCreationDTO:
+    uuid: str
+    name: str
+    is_template: bool
+    date_format: str
+    template_type_uuid: str
+    timezone: str
+    organization_id: int