+++ bothub/settings.py
@@ -118,7 +118,11 @@ env = environ.Env(
     AI_ODONTOLOGICAL_PLAN=(str, ""),
     AI_FINANCE_LEADS=(str, ""),
     AI_SENTIMENT_ANALYSIS=(str, ""),
-    CONNECT_WEBAPP_BASE_URL=(str, "http://localhost:8080")
+    CONNECT_WEBAPP_BASE_URL=(str, "http://localhost:8080"),
+    ZEROSHOT_BASE_NLP_URL=(str, ""),
+    FLOWS_TOKEN_ZEROSHOT=(str, ""),
+    ZEROSHOT_SUFFIX=(str, ""),
+    ZEROSHOT_TOKEN=(str, "")
 )
 
 # Build paths inside the project like this: os.path.join(BASE_DIR, ...)
@@ -163,6 +167,7 @@ INSTALLED_APPS = [
     "django_grpc_framework",
     "django_elasticsearch_dsl",
     "django_elasticsearch_dsl_drf",
+    "bothub.event_driven",
 ]
 
 MIDDLEWARE = [
@@ -677,3 +682,23 @@ RECOMMENDED_AIS = {
         AI_SENTIMENT_ANALYSIS,
     ]
 }
+
+# Event Driven Architecture configurations
+
+USE_EDA = env.bool("USE_EDA", default=False)
+
+if USE_EDA:
+    EDA_CONNECTION_BACKEND = "bothub.event_driven.connection.pymqp_connection.PyAMQPConnectionBackend"
+    EDA_CONSUMERS_HANDLE = "bothub.event_driven.handle.handle_consumers"
+
+    EDA_BROKER_HOST = env("EDA_BROKER_HOST", default="localhost")
+    EDA_VIRTUAL_HOST = env("EDA_VIRTUAL_HOST", default="/")
+    EDA_BROKER_PORT = env.int("EDA_BROKER_PORT", default=5672)
+    EDA_BROKER_USER = env("EDA_BROKER_USER", default="guest")
+    EDA_BROKER_PASSWORD = env("EDA_BROKER_PASSWORD", default="guest")
+    EDA_WAIT_TIME_RETRY = env("EDA_WAIT_TIME_RETRY", default=5)
+
+ZEROSHOT_BASE_NLP_URL = env.str("ZEROSHOT_BASE_NLP_URL")
+FLOWS_TOKEN_ZEROSHOT = env.str("FLOWS_TOKEN_ZEROSHOT")
+ZEROSHOT_SUFFIX = env.str("ZEROSHOT_SUFFIX")
+ZEROSHOT_TOKEN = env.str("ZEROSHOT_TOKEN")